\name{deltaT_ij}
\alias{deltaT_ij}
\docType{data}
\title{
%%   Matriz de coligação
}
\description{
%%  Matriz de coligação partidos relevantes
}
\usage{data("deltaT_ij")}
\format{
  A data frame with 11 observations on the following 11 variables.
  \describe{
    \item{\code{PPB}}{a numeric vector}
    \item{\code{PDT}}{a numeric vector}
    \item{\code{PT}}{a numeric vector}
    \item{\code{PTB}}{a numeric vector}
    \item{\code{PMDB}}{a numeric vector}
    \item{\code{PL}}{a numeric vector}
    \item{\code{PPS}}{a numeric vector}
    \item{\code{PFL}}{a numeric vector}
    \item{\code{PSB}}{a numeric vector}
    \item{\code{PSDB}}{a numeric vector}
    \item{\samp{PC do B}}{a numeric vector}
  }
}
\details{
%%  Matriz de coligação
}
\source{
%%  TSE
}
\references{
%%  Matriz de coligação
}
\examples{
data(deltaT_ij)
## maybe str(deltaT_ij) ; plot(deltaT_ij) ...
}
\keyword{datasets}
