\name{fit1}
\alias{fit1}
\docType{data}
\title{
%%   visualização
}
\description{
%%   visualização
}
\usage{data("fit1")}
\format{
  The format is:
List of 5
 $ points: num [1:11, 1:2] 19.8 -19.9 -32.2 18.9 15 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:11] "PPB" "PDT" "PT" "PTB" ...
  .. ..$ : NULL
 $ eig   : num [1:11] 4995 480 385 236 160 ...
 $ x     : NULL
 $ ac    : num 0
 $ GOF   : num [1:2] 0.828 0.828
}
\details{
%%   visualização
}
\source{
%%   visualização
}
\references{
%%   visualização
}
\examples{
data(fit1)
## maybe str(fit1) ; plot(fit1) ...
}
\keyword{datasets}
