\name{sigla_num}
\alias{sigla_num}
\docType{data}
\title{
%%   Partidos e Códigos
}
\description{
%%  Partidos e Códigos
}
\usage{data("sigla_num")}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{partido}}{a factor with levels \code{PC do B} \code{PDT} \code{PFL} \code{PL} \code{PMDB} \code{PMN} \code{PPB} \code{PPS} \code{PRONA} \code{PSB} \code{PSDB} \code{PSL} \code{PSTU} \code{PT} \code{PT do B} \code{PTB} \code{PV} \code{PRN} \code{PRP} \code{PRTB} \code{PSC} \code{PSDC} \code{PST} \code{PAN} \code{PSD} \code{PCB} \code{PGT} \code{PSN} \code{PTN} \code{PCO}}
    \item{\code{numero}}{a numeric vector}
  }
}
\details{
%%  Partidos e Códigos
}
\source{
%%  Partidos e Códigos
}
\references{
%%  Partidos e Códigos
}
\examples{
data(sigla_num)
## maybe str(sigla_num) ; plot(sigla_num) ...
}
\keyword{datasets}
